http: moving multiplexed active client
