+++ source/common/http/conn_pool_base.h
@@ -166,6 +166,51 @@ protected:
   const CreateClientFn client_fn_;
 };
 
+/**
+ * Active client base for HTTP/2 and HTTP/3
+ */
+class MultiplexedActiveClientBase : public CodecClientCallbacks,
+                                    public Http::ConnectionCallbacks,
+                                    public Envoy::Http::ActiveClient {
+public:
+  MultiplexedActiveClientBase(HttpConnPoolImplBase& parent, Stats::Counter& cx_total);
+  MultiplexedActiveClientBase(HttpConnPoolImplBase& parent, Stats::Counter& cx_total,
+                              Upstream::Host::CreateConnectionData& data);
+  ~MultiplexedActiveClientBase() override = default;
+
+  // ConnPoolImpl::ActiveClient
+  bool closingWithIncompleteStream() const override;
+  RequestEncoder& newStreamEncoder(ResponseDecoder& response_decoder) override;
+
+  // CodecClientCallbacks
+  void onStreamDestroy() override;
+  void onStreamReset(Http::StreamResetReason reason) override;
+
+  // Http::ConnectionCallbacks
+  void onGoAway(Http::GoAwayErrorCode error_code) override;
+  void onSettings(ReceivedSettings& settings) override;
+
+  // As this is called once when the stream is closed, it's a good place to
+  // update the counter as one stream has been "returned" and the negative
+  // capacity should be reduced.
+  bool hadNegativeDeltaOnStreamClosed() override {
+    int ret = negative_capacity_ != 0;
+    if (negative_capacity_ > 0) {
+      negative_capacity_--;
+    }
+    return ret;
+  }
+
+  uint64_t negative_capacity_{};
+
+protected:
+  MultiplexedActiveClientBase(Envoy::Http::HttpConnPoolImplBase& parent,
+                              Upstream::Host::CreateConnectionData& data, Stats::Counter& cx_total);
+
+private:
+  bool closed_with_active_rq_{};
+};
+
 } // namespace Http
 
 } // namespace Envoy
