+++ source/common/http/http3/BUILD
@@ -18,7 +18,6 @@ envoy_cc_library(
         "//include/envoy/upstream:upstream_interface",
         "//source/common/http:codec_client_lib",
         "//source/common/http:conn_pool_base_lib",
-        "//source/common/http/http2:conn_pool_lib",
     ],
 )
