+++ source/common/http/http3/conn_pool.h
@@ -5,7 +5,6 @@
 #include "envoy/upstream/upstream.h"
 
 #include "common/http/codec_client.h"
-#include "common/http/http2/conn_pool.h"
 
 namespace Envoy {
 namespace Http {
